requirements.txt
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
cssbeautifier==1.14.11
djlint==1.34.0
EditorConfig==0.12.3
Flask==2.3.2
Flask-Assets==2.1.0
Flask-Session==0.5.0
greenlet==2.0.2
gunicorn==21.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsbeautifier==1.14.11
json5==0.9.14
MarkupSafe==2.1.3
packaging==23.2
pathspec==0.11.2
pip==23.0.1
psycopg2-binary==2.9.6
pycparser==2.21
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
setuptools==66.1.1
six==1.16.0
SQLAlchemy==2.0.16
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
webassets==2.0
Werkzeug==2.3.7
wheel==0.42.0
